==> bulk_state_of_the_art/notebooks/01_preprocessing/02_merge_gdsc_datasets.ipynb <==
import polars as pl

# Load GDSC1 and GDSC2
gdsc = pl.read_parquet("../../data/processed/gdsc_with_split_drug_ids.parquet")

# Display first few rows
print(gdsc.head())

# Essential columns to keep
columns_to_keep = ["SANGER_MODEL_ID", "DRUG_ID", "LN_IC50"]

# Drop unnecessary columns
gdsc_cleaned = gdsc[columns_to_keep]

print(f"🔹 Cleaned dataset shape: {gdsc_cleaned.shape}")


# Check for missing values
missing_ln_ic50 = gdsc_cleaned.filter(pl.col("LN_IC50").is_null()).shape[0]
print(f"\n📊 Missing Values in LN_IC50: {missing_ln_ic50}")

if missing_ln_ic50 > 0:
    # Fill NaNs with group mean
    gdsc_cleaned = (
        gdsc_cleaned
        .with_columns([
            pl.col("LN_IC50").fill_null(
                pl.col("LN_IC50").mean().over(["SANGER_MODEL_ID", "DRUG_ID"])
            )
        ])
        .with_columns([
            pl.col("LN_IC50").fill_null(0)  # Fill any remaining nulls with 0
        ])
    )


gdsc_cleaned.write_parquet("../../data/processed/gdsc_final_cleaned.parquet")
print("✅ Final merged and cleaned GDSC dataset saved!")
print(f"📌 Final dataset shape: {gdsc_cleaned.shape}")
